third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-lifetime-mediakeysession-release-noreference.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test MediaKeySession lifetime after release() without references</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
            // Since MediaKeySession (and MediaKeys) are SuspendableObjects,
            // we can determine when they are garbage collected.
            // MediaKeySessions remain as long as:
            //   JavaScript has a reference to it
            //   OR (MediaKeys is around
            //       AND the session has not received a close() event)
            async_test(function(test)
            {
                var initDataType;
                var initData;
                var startingSuspendableObjectCount = window.internals.suspendableObjectCount(document);

                function numSuspendableObjectsCreated()
                {
                    return window.internals.suspendableObjectCount(document) - startingSuspendableObjectCount;
                }

                // Create 2 sessions.
                var mediaKeys;
                var mediaKeySession1;
                var mediaKeySession2;

                navigator.requestMediaKeySystemAccess('org.w3.clearkey', getSimpleConfiguration()).then(function(access) {
                    initDataType = access.getConfiguration().initDataTypes[0];
                    initData = getInitData(initDataType);
                    return access.createMediaKeys();
                }).then(function(result) {
                    mediaKeys = result;

                    // Verify MediaKeys is an SuspendableObject.
                    // In non-Oilpan, numSuspendableObjectsCreate() == 1.
                    // In Oilpan, numSuspendableObjectsCreate() <= 4.
                    // (1 MediaKeys,
                    //  1 MediaKeysInitializer and
                    //  1 MediaKeySystemAccessInitializer (navigator.requestMediaKeySystemAccess() use above),
                    //  1 MediaKeySystemAccessInitializer (isInitDataSupported() (via getSupportedInitDataType())))
                    assert_between_inclusive(numSuspendableObjectsCreated(), 1, 4, 'MediaKeys.create()');

                    mediaKeySession1 = mediaKeys.createSession();
                    return mediaKeySession1.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_true(mediaKeySession1.sessionId && mediaKeySession1.sessionId.length > 0);

                    // Should be 1 MediaKeys + 1 MediaKeySession.
                    // In non-Oilpan, numSuspendableObjectsCreate() == 2.
                    // In Oilpan, numSuspendableObjectsCreate() <= 6.
                    // (1 MediaKeys,
                    //  1 MediaKeysInitializer and
                    //  2 MediaKeySystemAccessInitializer,
                    //  1 ContentDecryptionModuleResultPromise and
                    //  1 MediaKeySession).
                    assert_between_inclusive(numSuspendableObjectsCreated(), 2, 6, 'MediaKeys.createSession(1)');

                    mediaKeySession2 = mediaKeys.createSession();
                    return mediaKeySession2.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_true(mediaKeySession2.sessionId && mediaKeySession2.sessionId.length > 0);

                    // Should be 1 MediaKeys + 2 MediaKeySessions.
                    // In non-Oilpan, numSuspendableObjectsCreate() == 3.
                    // In Oilpan, numSuspendableObjectsCreate() <= 8.
                    // (1 MediaKeys,
                    //  1 MediaKeysInitializer and
                    //  2 MediaKeySystemAccessInitializers,
                    //  2 ContentDecryptionModuleResultPromise and
                    //  2 MediaKeySession).
                    assert_between_inclusive(numSuspendableObjectsCreated(), 3, 8, 'mediaKeys.createSession(2)');
                }).then(function(result) {
                    // Run gc(). All sessions should remain as we have a
                    // reference to each one.
                    return createGCPromise();
                }).then(function(result) {
                    // Should be just 1 MediaKeys + 2 MediaKeySessions.
                    // In non-Oilpan, there is also something from createGCPromise().
                    assert_between_inclusive(numSuspendableObjectsCreated(), 3, 4, 'After gc()');

                    // Close the sessions. Once the close() event is received,
                    // they should get garbage collected as there are no JS
                    // references to them.
                    var promise = mediaKeySession1.close();
                    mediaKeySession1 = null;
                    return promise;
                }).then(function(result) {
                    // Give time so that the close event can be processed by
                    // MediaKeySession.
                    return delayToAllowEventProcessingPromise();
                }).then(function(result) {
                    return createGCPromise();
                }).then(function(result) {
                    // Only MediaKeys + mediaKeySession2 should remain.
                    // In non-Oilpan, there is also something from createGCPromise().
                    assert_between_inclusive(numSuspendableObjectsCreated(), 2, 3, 'mediaKeySession1 not collected');

                    var promise = mediaKeySession2.close();
                    mediaKeySession2 = null;
                    return promise;
                }).then(function(result) {
                    // Provide time for the mediaKeySession2 close event to be
                    // handled.
                    return delayToAllowEventProcessingPromise();
                }).then(function(result) {
                    return createGCPromise();
                }).then(function(result) {
                    // Only MediaKeys should remain.
                    // In non-Oilpan, there is also something from createGCPromise().
                    assert_between_inclusive(numSuspendableObjectsCreated(), 1, 2, 'mediaKeySession2 not collected');

                    assert_not_equals(mediaKeys, null);
                    test.done();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'MediaKeySession lifetime after release() without references');
        </script>
    </body>
</html>
